+++ agenta-web/src/components/EvaluationTable/ABTestingEvaluationTable.tsx
@@ -18,7 +18,7 @@ import {
     fetchEvaluationResults,
     updateEvaluation,
 } from "@/services/human-evaluations"
-import {callVariant} from "@/lib/services/api"
+import {callVariant} from "@/services"
 import {useVariants} from "@/lib/hooks/useVariant"
 import {useRouter} from "next/router"
 import {EvaluationFlow} from "@/lib/enums"
