+++ agenta-web/src/components/HumanEvaluationModal/HumanEvaluationModal.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from "react"
 import {GenericObject, JSSTheme, Parameter, Variant} from "@/lib/Types"
-import {fetchVariants} from "@/lib/services/api"
+import {fetchVariants} from "@/services"
 import {createNewEvaluation} from "@/services/human-evaluations"
 import {isDemo} from "@/lib/helpers/utils"
 import {Button, Col, Dropdown, MenuProps, Modal, ModalProps, Row, Spin, message} from "antd"
