+++ agenta-web/src/components/Layout/Layout.tsx
@@ -20,7 +20,7 @@ import {createUseStyles} from "react-jss"
 import NoSSRWrapper from "../NoSSRWrapper/NoSSRWrapper"
 import {ErrorBoundary} from "react-error-boundary"
 import ErrorFallback from "./ErrorFallback"
-import {fetchData} from "@/lib/services/api"
+import {fetchData} from "@/services"
 import {useAppsData} from "@/contexts/app.context"
 import {useRouter} from "next/router"
 import Image from "next/image"
