+++ agenta-web/src/pages/apps/[app_id]/annotations/single_model_test/[evaluation_id]/index.tsx
@@ -3,7 +3,7 @@ import {loadEvaluation, loadEvaluationsScenarios} from "@/services/human-evaluat
 import {loadTestset} from "@/services/testsets"
 import {useRouter} from "next/router"
 import {useEffect, useState} from "react"
-import {fetchVariants} from "@/lib/services/api"
+import {fetchVariants} from "@/services"
 import {getTestsetChatColumn} from "@/lib/helpers/testset"
 import SingleModelEvaluationTable from "@/components/EvaluationTable/SingleModelEvaluationTable"
